moving station class to a new file
